telegram/requirements.txt
python-telegram-bot==21.0.1
requests==2.31.0
fpdf==1.7.2
pillow==10.2.0